Add a sort command for task list
Tasks are ordered by the order which they are added to the task list.
This is messy and hard for the user to see which tasks are of higher
priority in terms of due date and start time.

Event and Deadline tasks are sorted by chronological order, while
Todo tasks are sorted by alphabetical order, so the user can easily
view which tasks should be completed first.